# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinBank Microservices API
  version: 0.13.0
  description: Curated aggregate contract for the account, payment, and transaction services.
...
paths:
  /accounts:
    servers:
      - url: http://localhost:8081
        description: Account Service
    get:
      operationId: listAccounts
      tags: [Accounts]
      summary: List accounts
      description: >-
        Returns at most limit accounts in ascending id order. Pass the final id
        from a rel="next" Link target back as afterId to continue.
      parameters:
        - name: afterId
          in: query
          required: false
          description: Return rows whose id is strictly greater than this exclusive cursor.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          required: false
          description: Maximum rows returned; defaults to 50 and cannot exceed 100.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Bounded account page returned in ascending id order
          headers:
            Link:
              description: >-
                Present only when another row was observed. The rel="next"
                target uses the final returned id as the next exclusive cursor.
              schema:
                type: string
              example: '</accounts?afterId=202&limit=50>; rel="next"'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
              examples:
                populated:
                  summary: Synthetic account list
                  value:
                    - id: 101
                      customerName: Demo Sender
                      balance: 5000.00
                    - id: 202
                      customerName: Demo Recipient
                      balance: 2500.00
                empty:
                  summary: No accounts stored
                  value: []
        "400":
          description: Pagination cursor or limit failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidPagination:
                  summary: Safe pagination failure without rejected values
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /accounts
    post:
      operationId: createAccount
      tags: [Accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
              examples:
                created:
                  summary: Synthetic newly created account
                  value:
                    id: 101
                    customerName: Demo Customer
                    balance: 5000.00
        "400":
          description: Customer name or opening balance failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidRequest:
                  summary: Safe validation failure without rejected values
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /accounts
  /accounts/{id}:
    servers:
      - url: http://localhost:8081
        description: Account Service
    get:
      operationId: getAccountById
      tags: [Accounts]
      summary: Get an account
      description: Returns one account identified by a positive generated id.
      parameters:
        - name: id
          in: path
          required: true
          description: Positive account identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
              examples:
                found:
                  summary: Synthetic account lookup result
                  value:
                    id: 101
                    customerName: Demo Customer
                    balance: 5000.00
        "400":
          description: Account id failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidId:
                  summary: Safe path validation failure without the rejected value
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /accounts
        "404":
          description: Account was not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicNotFoundProblem"
              examples:
                notFound:
                  summary: Safe account lookup miss without the requested identifier
                  value:
                    type: urn:finbank:problem:resource-not-found
                    title: Resource not found
                    status: 404
                    detail: The requested resource was not found.
                    instance: /accounts
  /payments:
    servers:
      - url: http://localhost:8082
        description: Payment Service
    get:
      operationId: listPayments
      tags: [Payments]
      summary: List payments
      description: >-
        Returns at most limit payments in ascending id order. Pass the final id
        from a rel="next" Link target back as afterId to continue.
      parameters:
        - name: afterId
          in: query
          required: false
          description: Return rows whose id is strictly greater than this exclusive cursor.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          required: false
          description: Maximum rows returned; defaults to 50 and cannot exceed 100.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Bounded payment page returned in ascending id order
          headers:
            Link:
              description: >-
                Present only when another row was observed. The rel="next"
                target uses the final returned id as the next exclusive cursor.
              schema:
                type: string
              example: '</payments?afterId=201&limit=50>; rel="next"'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
              examples:
                populated:
                  summary: Synthetic payment after successful publication
                  value:
                    - id: 201
                      fromAccount: 101
                      toAccount: 202
                      amount: 750.00
                      status: PUBLISHED
                empty:
                  summary: No payments stored
                  value: []
        "400":
          description: Pagination cursor or limit failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidPagination:
                  summary: Safe pagination failure without rejected values
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /payments
    post:
      operationId: createPayment
      tags: [Payments]
      summary: Create a payment and queue its event for publication
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Opaque case-sensitive key for safely replaying this request.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '^[!-~]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "200":
          description: Payment created with durable event publication queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
              examples:
                created:
                  summary: Synthetic payment with publication queued
                  value:
                    id: 201
                    fromAccount: 101
                    toAccount: 202
                    amount: 750.00
                    status: CREATED
        "400":
          description: Idempotency key or payment request failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidRequest:
                  summary: Safe validation failure without request values
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /payments
        "409":
          description: Idempotency key was already used for a different payment
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PaymentIdempotencyConflictProblem"
              examples:
                conflict:
                  summary: Safe conflict without the idempotency key or payment identity
                  value:
                    type: urn:finbank:problem:idempotency-key-conflict
                    title: Idempotency key conflict
                    status: 409
                    detail: The Idempotency-Key is already associated with a different payment request.
                    instance: /payments
  /payments/{id}:
    servers:
      - url: http://localhost:8082
        description: Payment Service
    get:
      operationId: getPaymentById
      tags: [Payments]
      summary: Get a payment
      description: Returns one payment identified by a positive generated id.
      parameters:
        - name: id
          in: path
          required: true
          description: Positive payment identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
              examples:
                found:
                  summary: Synthetic payment lookup result
                  value:
                    id: 201
                    fromAccount: 101
                    toAccount: 202
                    amount: 750.00
                    status: PUBLISHED
        "400":
          description: Payment id failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidId:
                  summary: Safe path validation failure without the rejected value
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /payments
        "404":
          description: Payment was not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicNotFoundProblem"
              examples:
                notFound:
                  summary: Safe payment lookup miss without the requested identifier
                  value:
                    type: urn:finbank:problem:resource-not-found
                    title: Resource not found
                    status: 404
                    detail: The requested resource was not found.
                    instance: /payments
...
  /transactions/account/{id}:
    servers:
      - url: http://localhost:8083
        description: Transaction Service
    get:
      operationId: listTransactionsByAccount
      tags: [Transactions]
      summary: List transactions involving one account
      description: >-
        Returns at most limit transactions involving the requested account in
        ascending id order. Pass the final id from a rel="next" Link target back
        as afterId to continue for the same account.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: afterId
          in: query
          required: false
          description: Return rows whose id is strictly greater than this exclusive cursor.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          required: false
          description: Maximum rows returned; defaults to 50 and cannot exceed 100.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Bounded transaction account-history page returned in ascending id order
          headers:
            Link:
              description: >-
                Present only when another row was observed. The rel="next"
                target preserves the requested account and uses the final
                returned id as the next exclusive cursor.
              schema:
                type: string
              example: '</transactions/account/101?afterId=301&limit=50>; rel="next"'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
              examples:
                matchingAccount:
                  summary: Synthetic ledger entry involving the requested account
                  value:
                    - id: 301
                      paymentId: 201
                      fromAccount: 101
                      toAccount: 202
                      amount: 750.00
                      createdAt: "2026-01-15T10:30:00Z"
                      status: COMPLETED
                empty:
                  summary: No transactions involve the requested account
                  value: []
        "400":
          description: Account id, pagination cursor, or limit failed validation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/PublicValidationProblem"
              examples:
                invalidRequestValue:
                  summary: Safe path or query validation failure without rejected values
                  value:
                    type: urn:finbank:problem:validation-failed
                    title: Request validation failed
                    status: 400
                    detail: One or more request values are invalid.
                    instance: /transactions/account
components:
  securitySchemes:
    recoveryOperator:
      type: http
      scheme: basic
      description: >-
        Local Basic operator credential selected only in basic mode and supplied
        through runtime configuration. Activation requires either the direct TLS
        connector or the explicit trusted-proxy transport mode.
    handoffInspectionOperator:
      type: http
      scheme: basic
      description: >-
        Configured local Basic operator credential with
        PAYMENT_OUTBOX_HANDOFF_INSPECTION authority. The current operator also
        has recovery authority; JWT mode can separate the roles through exact
        scopes.
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        External RS256 JWT selected instead of Basic authentication. Tokens are
        issuer-, audience-, and subject-validated, require iat and exp, and honor
        nbf when present with 60 seconds of clock-skew tolerance. Standard scope
        and scp claims are accepted. Only the documented payment.outbox.recovery
        and payment.outbox.handoff-inspection scopes map to operator authorities.
  schemas:
    Account:
      type: object
      description: >-
        Explicit public account response projection. Persistence entity fields
        are not serialized directly.
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 101
        customerName:
          type: string
          description: Customer display name accepted by the account creation API.
          example: Demo Customer
        balance:
          type: number
          format: decimal
          description: Current account balance represented as a decimal number.
          example: 5000.00
    CreateAccountRequest:
      type: object
      description: Client input for account creation; the service assigns the account id.
      required: [customerName, balance]
      properties:
        customerName:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          description: Must contain at least one non-whitespace character.
          example: Demo Customer
        balance:
          type: number
          format: decimal
          minimum: 0
          example: 5000.00
    Payment:
      type: object
      description: >-
        Explicit public payment response projection. Persistence entity fields
        are not serialized directly, and the internal idempotency-key digest is
        not part of this contract.
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 201
        fromAccount:
          type: integer
          format: int64
          minimum: 1
          example: 101
        toAccount:
          type: integer
          format: int64
          minimum: 1
          example: 202
        amount:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.01
          example: 750.00
        status:
          type: string
          readOnly: true
          enum: [CREATED, PENDING_RETRY, PUBLISHED, PUBLISH_EXHAUSTED]
          description: >-
            Server-owned publication lifecycle state; a successful create starts
            at CREATED and the outbox relay advances it later.
          example: CREATED
    CreatePaymentRequest:
      type: object
      description: Client input for a payment; the service assigns the payment id and status.
      required: [fromAccount, toAccount, amount]
      properties:
        fromAccount:
          type: integer
          format: int64
          minimum: 1
          example: 1
        toAccount:
          type: integer
          format: int64
          minimum: 1
          description: Must identify an account different from `fromAccount`.
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.01
          example: 750.00
...
    PublicValidationProblem:
      type: object
      description: >-
        Stable public validation failure. Rejected values, field errors,
        framework messages, and exception details are intentionally omitted.
      additionalProperties: false
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          enum: [urn:finbank:problem:validation-failed]
        title:
          type: string
          enum: [Request validation failed]
        status:
          type: integer
          format: int32
          enum: [400]
        detail:
          type: string
          enum: [One or more request values are invalid.]
        instance:
          type: string
          enum: [/accounts, /payments, /transactions, /transactions/account]
    PaymentIdempotencyConflictProblem:
      type: object
      description: >-
        Stable payment idempotency conflict that never returns the supplied key,
        its digest, the matching payment, or request-comparison details.
      additionalProperties: false
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          enum: [urn:finbank:problem:idempotency-key-conflict]
        title:
          type: string
          enum: [Idempotency key conflict]
        status:
          type: integer
          format: int32
          enum: [409]
        detail:
          type: string
          enum: [The Idempotency-Key is already associated with a different payment request.]
        instance:
          type: string
          enum: [/payments]
    PublicNotFoundProblem:
      type: object
      description: >-
        Stable public lookup miss. Requested identifiers, persistence details,
        framework messages, and exception details are intentionally omitted.
      additionalProperties: false
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          enum: [urn:finbank:problem:resource-not-found]
        title:
          type: string
          enum: [Resource not found]
        status:
          type: integer
          format: int32
          enum: [404]
        detail:
          type: string
          enum: [The requested resource was not found.]
        instance:
          type: string
          enum: [/accounts, /payments]
    Transaction:
      type: object
      description: >-
        Explicit public transaction response projection for a read-only ledger
        entry derived from a validated payment event. Persistence entity fields
        are not serialized directly.
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 301
        paymentId:
          type: integer
          format: int64
          minimum: 1
          description: Payment event identity used to make ledger writes idempotent.
          example: 201
        fromAccount:
          type: integer
          format: int64
          minimum: 1
          example: 101
        toAccount:
          type: integer
          format: int64
          minimum: 1
          example: 202
        amount:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.01
          example: 750.00
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-01-15T10:30:00Z"
        status:
          type: string
          readOnly: true
          description: Ledger processing state assigned by the transaction service.
          example: COMPLETED